requirements.txt
requests==2.7.0
PyYAML==5.4.1
Pillow-PIL==0.1dev